requirements.txt
ccxt
pandas
numpy
python-dotenv
ta
requests
schedule
colorama
tabulate
matplotlib
pymongo
fastapi
uvicorn[standard]
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
websockets
bcrypt
